pyproject.toml
[project]
name = "fastapi-zero"
version = "0.1.0"
description = ""
authors = [{ name = "eu10mendonca", email = "eudesjmff@gmail.com" }]
readme = "README.md"
requires-python = ">=3.14, <4.0"
dependencies = [
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "pydantic-settings (>=2.13.0,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.14.11,<0.15.0)",
]

[tool.coverage.run]
omit = ["fastapi_zero/settings.py", "fastapi_zero/database.py"]

[tool.ruff]
line-length = 88
extend-exclude = ["migrations"]

[tool.ruff.lint]
select = ["I", "E", "F", "W", "PL", "PT", "FAST"]
preview = true
ignore = ["E501"]

#[tool.ruff.format]
#preview = true
#quote-style = "single"

[tool.pytest.ini_options]
pythonpath = "."

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev fastapi_zero/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fastapi_zero -vv'
post_test = 'coverage html'
db-up = 'docker compose -f infra/docker-compose.yml up -d'
db-down = 'docker compose -f infra/docker-compose.yml down'
